requirements.txt
pydantic-settings
python-dotenv

python-docx
pandas

chromadb
sentence-transformers

ollama

gradio